Use email address table to confirm email during sign up
**Why**: To stop reading from the `email` column in the user table so we can eventually drop it.
